feat(pkgs): add lima, colima and dnsmasq
Trying lima and colima together with dnsmasq as alternative for a local
k8s stack. Lima is the same foundation that is part of Rancher Desktop.
Colima is "Containers on Lima" and provides the routines to provision
a k3s on the lima vm.

I like to use dnsmasq as a local wildcard dns server to have valid
dns local domain names for ingress tests.

Kindly the MacOS provides an option to define different dns servers
for different zones, see `scutil --dns` for more details.